Adding the tracing for sequences using the pranet-child hierarchy.
Sequence action starts its own span and component actions are then
created as child spans so the tracing reflects what is actually
happening in the event based system, including the simple causality
by following the sequences.

It looks like the activationId or cause are good candidates for the
correlation id among the spans, however, in order to preserve the open-tracing
ideas with tracer neutrality (we don't like vendor lock-ins, right?) the
mutable Map[String,String] needs to be propagated from the parent spans
(started in the controller component) to the child spans (started in the
invokers). To do that, I had to add a new attribute the to
ActivationMessage and also change couple of signatures in the
SequenceActions.scala.

=> No vendor lock-in on the tracing solution = Profit!